#206 Move transaction state machine into separate module
