+++ src/producer/transactionStates.js
@@ -1,0 +1,7 @@
+module.exports = {
+  UNINITIALIZED: 'UNINITIALIZED',
+  READY: 'READY',
+  TRANSACTING: 'TRANSACTING',
+  COMMITTING: 'COMMITTING',
+  ABORTING: 'COMMITTING',
+}
